Initial server-side API implementation Features a basic account system and exposes an endpoint for publishing levels, though currently only a test button after logging in does the publishing.
Since this is quite a large commit and this is a peer-reviewed course, I feel like I should probably justify its size a bit, since there is a considerable chance someone new to Git might be reviewing this and be confused by such a decision.

The main use case of having small, self-contained commits is for reproducibility; if something breaks, you can see what the probable cause is and then just pick out that one commit to figure out the culprit. If the project has multiple contributors, it's also much easier for others to understand one's changes when they are small self-contained patches rather than megalomaniac overhauls.

However, in this case, this is a solo project that is still at a very early stage; there is essentially nothing to break, since there is nothing to use, and there is no one else to trip up with commits like this. Therefore, there exists no need to split this commit into smaller parts, since those smaller parts would not help anyone. Splitting a commit into many smaller parts would just clog up the commit history with tiny commits that don't really help an observer form a better understanding of the project's evolution, since at such an early stage like this there's not much of interest to look at.

It's also usually preferable that the project at the point of a commit is at least locally runnable, if not usable or test-passing. You don't want to commit incomplete code full of syntax errors, as that just results in a commit that helps no one: if you can't even build it, you can't return to that point in time, so the commit can't possibly do its job of being a snapshot of an earlier working version.

This is why even on non-personal, non-early-stage projects, huge overhaul commits are sometimes warranted, if it really can't be broken down to smaller parts that A) make sense and B) can be built into at least a runnable application.

Thanks for reading my essay :3